moved get_subfield_from_code to utils.py
